+++ commands/GemstonesOnFaceAtCurve/GemstonesOnFaceAtCurve.py
@@ -22,6 +22,7 @@ _sizeStepValueInput: adsk.core.ValueCommandInput = None
 _targetGapValueInput: adsk.core.ValueCommandInput = None
 _flipValueInput: adsk.core.BoolValueCommandInput = None
 _flipDirectionValueInput: adsk.core.BoolValueCommandInput = None
+_uniformDistributionValueInput: adsk.core.BoolValueCommandInput = None
 _absoluteDepthOffsetValueInput: adsk.core.ValueCommandInput = None
 _relativeDepthOffsetValueInput: adsk.core.ValueCommandInput = None
 _nonlinearValueInput: adsk.core.BoolValueCommandInput = None
@@ -34,9 +35,7 @@ _isRolledForEdit: bool = False
 
 _handlers = []
 
-COMMAND_ID = strings.PREFIX + strings.GEMSTONES_ON_FACE_AT_CURVE
-CREATE_COMMAND_ID = COMMAND_ID + 'Create'
-EDIT_COMMAND_ID = COMMAND_ID + 'Edit'
+COMMAND_ID, CREATE_COMMAND_ID, EDIT_COMMAND_ID = strings.getCommandIds(strings.GEMSTONES_ON_FACE_AT_CURVE)
 
 createCommandInputDef = strings.InputDef(CREATE_COMMAND_ID, 'Gemstones at Curves', 'Creates gemstones at selected curves on a face.')
 editCommandInputDef = strings.InputDef(EDIT_COMMAND_ID, 'Edit Gemstones', 'Edits the parameters of existing gemstones.')
@@ -99,6 +98,12 @@ flipDirectionInputDef = strings.InputDef(
     'flipDirection',
     'Flip Direction',
     "Flip gemstone placement direction.\nStarts placing gemstones from the opposite end of the curve."
+    )
+
+uniformDistributionInputDef = strings.InputDef(
+    'uniformDistribution',
+    'Uniform Distribution',
+    "Distribute gemstones uniformly along the curve.\nEnsures gemstones fill the entire available length\nfrom start offset to end offset without gaps at the ends."
     )
 
 absoluteDepthOffsetInputDef = strings.InputDef(
@@ -207,7 +212,8 @@ class CreateCommandCreatedHandler(adsk.core.CommandCreatedEventHandler):
 
             global _faceSelectionInput, _curveSelectionInput, _startOffsetValueInput, _endOffsetValueInput
             global _startSizeValueInput, _endSizeValueInput, _sizeStepValueInput, _targetGapValueInput
-            global _flipValueInput, _flipDirectionValueInput, _absoluteDepthOffsetValueInput, _relativeDepthOffsetValueInput
+            global _flipValueInput, _flipDirectionValueInput, _uniformDistributionValueInput
+            global _absoluteDepthOffsetValueInput, _relativeDepthOffsetValueInput
             global _nonlinearValueInput, _nonlinearSizeValueInput, _nonlinearPositionValueInput
 
             _curveSelectionInput = inputs.addSelectionInput(selectCurveInputDef.id, selectCurveInputDef.name, selectCurveInputDef.tooltip)
@@ -227,6 +233,10 @@ class CreateCommandCreatedHandler(adsk.core.CommandCreatedEventHandler):
             flipDirection = False
             _flipDirectionValueInput = inputs.addBoolValueInput(flipDirectionInputDef.id, flipDirectionInputDef.name, True, '', flipDirection)
             _flipDirectionValueInput.tooltip = flipDirectionInputDef.tooltip
+
+            uniformDistribution = False
+            _uniformDistributionValueInput = inputs.addBoolValueInput(uniformDistributionInputDef.id, uniformDistributionInputDef.name, True, '', uniformDistribution)
+            _uniformDistributionValueInput.tooltip = uniformDistributionInputDef.tooltip
             
             startOffset = adsk.core.ValueInput.createByReal(0.0)
             _startOffsetValueInput = inputs.addValueInput(startOffsetInputDef.id, startOffsetInputDef.name, defaultLengthUnits, startOffset)
@@ -329,6 +339,7 @@ class EditCommandCreatedHandler(adsk.core.CommandCreatedEventHandler):
 
             global _startOffsetValueInput, _endOffsetValueInput, _startSizeValueInput, _endSizeValueInput
             global _sizeStepValueInput, _targetGapValueInput, _flipValueInput, _flipDirectionValueInput
+            global _uniformDistributionValueInput
             global _absoluteDepthOffsetValueInput, _relativeDepthOffsetValueInput
             global _nonlinearValueInput, _nonlinearSizeValueInput, _nonlinearPositionValueInput
 
@@ -355,6 +366,14 @@ class EditCommandCreatedHandler(adsk.core.CommandCreatedEventHandler):
                 flipDirection = False
             _flipDirectionValueInput = inputs.addBoolValueInput(flipDirectionInputDef.id, flipDirectionInputDef.name, True, '', flipDirection)
             _flipDirectionValueInput.tooltip = flipDirectionInputDef.tooltip
+
+            try:
+                uniformDistributionParam = params.itemById(uniformDistributionInputDef.id)
+                uniformDistribution = uniformDistributionParam.expression.lower() == 'true'
+            except:
+                uniformDistribution = False
+            _uniformDistributionValueInput = inputs.addBoolValueInput(uniformDistributionInputDef.id, uniformDistributionInputDef.name, True, '', uniformDistribution)
+            _uniformDistributionValueInput.tooltip = uniformDistributionInputDef.tooltip
 
             try:
                 startOffsetParam = params.itemById(startOffsetInputDef.id)
@@ -527,6 +546,7 @@ class ValidateInputsHandler(adsk.core.ValidateInputsEventHandler):
                          _startSizeValueInput.isValidExpression, _endSizeValueInput.isValidExpression,
                          _sizeStepValueInput.isValidExpression, _targetGapValueInput.isValidExpression,
                          _flipValueInput.isValid, _flipDirectionValueInput.isValid,
+                         _uniformDistributionValueInput.isValid,
                          _absoluteDepthOffsetValueInput.isValidExpression,
                          _relativeDepthOffsetValueInput.isValidExpression,
                          _nonlinearValueInput.isValid,
@@ -600,13 +620,14 @@ class ExecutePreviewHandler(adsk.core.CommandEventHandler):
             targetGap = _targetGapValueInput.value
             flip = _flipValueInput.value
             flipDirection = _flipDirectionValueInput.value
+            uniformDistribution = _uniformDistributionValueInput.value
             absoluteDepthOffset = _absoluteDepthOffsetValueInput.value
             relativeDepthOffset = _relativeDepthOffsetValueInput.value
             nonlinear = _nonlinearValueInput.value
             nonlinearSize = _nonlinearSizeValueInput.value
             nonlinearPosition = _nonlinearPositionValueInput.value
 
-            pointsAndSizes = calculatePointsAndSizesAlongCurve(curve, startOffset, endOffset, startSize, endSize, sizeStep, targetGap, flipDirection, nonlinear, nonlinearSize, nonlinearPosition)
+            pointsAndSizes = calculatePointsAndSizesAlongCurve(curve, startOffset, endOffset, startSize, endSize, sizeStep, targetGap, flipDirection, uniformDistribution, nonlinear, nonlinearSize, nonlinearPosition)
             if len(pointsAndSizes) == 0:
                 return
 
@@ -640,6 +661,7 @@ class CreateExecuteHandler(adsk.core.CommandEventHandler):
         try:
             eventArgs = adsk.core.CommandEventArgs.cast(args)        
 
+            comp: adsk.fusion.Component = None
             face = _faceSelectionInput.selection(0).entity
             curveEntity = _curveSelectionInput.selection(0).entity
             if face.objectType == adsk.fusion.ConstructionPlane.classType():
@@ -655,7 +677,7 @@ class CreateExecuteHandler(adsk.core.CommandEventHandler):
             pointsAndSizes = calculatePointsAndSizesAlongCurve(curve, _startOffsetValueInput.value, _endOffsetValueInput.value,
                                                                _startSizeValueInput.value, _endSizeValueInput.value,
                                                                _sizeStepValueInput.value, _targetGapValueInput.value,
-                                                               _flipDirectionValueInput.value,
+                                                               _flipDirectionValueInput.value, _uniformDistributionValueInput.value,
                                                                _nonlinearValueInput.value, _nonlinearSizeValueInput.value, _nonlinearPositionValueInput.value)
             if len(pointsAndSizes) == 0:
                 eventArgs.executeFailed = True
@@ -710,6 +732,9 @@ class CreateExecuteHandler(adsk.core.CommandEventHandler):
 
             flipDirectionInput = adsk.core.ValueInput.createByString(str(_flipDirectionValueInput.value).lower())
             customFeatureInput.addCustomParameter(flipDirectionInputDef.id, flipDirectionInputDef.name, flipDirectionInput, '', True)
+
+            uniformDistributionInput = adsk.core.ValueInput.createByString(str(_uniformDistributionValueInput.value).lower())
+            customFeatureInput.addCustomParameter(uniformDistributionInputDef.id, uniformDistributionInputDef.name, uniformDistributionInput, '', True)
 
             absoluteDepthOffsetInput = adsk.core.ValueInput.createByString(_absoluteDepthOffsetValueInput.expression)
             customFeatureInput.addCustomParameter(absoluteDepthOffsetInputDef.id, absoluteDepthOffsetInputDef.name, absoluteDepthOffsetInput,
@@ -809,31 +834,20 @@ class EditExecuteHandler(adsk.core.CommandEventHandler):
             _editedCustomFeature.dependencies.add('face', faceEntity)
             _editedCustomFeature.dependencies.add('curve', curveEntity)
 
+            _editedCustomFeature.parameters.itemById(flipDirectionInputDef.id).expression = str(_flipDirectionValueInput.value).lower()
+            _editedCustomFeature.parameters.itemById(uniformDistributionInputDef.id).expression = str(_uniformDistributionValueInput.value).lower()
             _editedCustomFeature.parameters.itemById(startOffsetInputDef.id).expression = _startOffsetValueInput.expression
             _editedCustomFeature.parameters.itemById(endOffsetInputDef.id).expression = _endOffsetValueInput.expression
             _editedCustomFeature.parameters.itemById(startSizeInputDef.id).expression = _startSizeValueInput.expression
             _editedCustomFeature.parameters.itemById(endSizeInputDef.id).expression = _endSizeValueInput.expression
+            _editedCustomFeature.parameters.itemById(nonlinearInputDef.id).expression = str(_nonlinearValueInput.value).lower()
+            _editedCustomFeature.parameters.itemById(nonlinearSizeInputDef.id).expression = _nonlinearSizeValueInput.expression
+            _editedCustomFeature.parameters.itemById(nonlinearPositionInputDef.id).expression = _nonlinearPositionValueInput.expression
             _editedCustomFeature.parameters.itemById(sizeStepInputDef.id).expression = _sizeStepValueInput.expression
             _editedCustomFeature.parameters.itemById(targetGapInputDef.id).expression = _targetGapValueInput.expression
             _editedCustomFeature.parameters.itemById(flipInputDef.id).expression = str(_flipValueInput.value).lower()
-            _editedCustomFeature.parameters.itemById(flipDirectionInputDef.id).expression = str(_flipDirectionValueInput.value).lower()
             _editedCustomFeature.parameters.itemById(absoluteDepthOffsetInputDef.id).expression = _absoluteDepthOffsetValueInput.expression
             _editedCustomFeature.parameters.itemById(relativeDepthOffsetInputDef.id).expression = _relativeDepthOffsetValueInput.expression
-
-            try:
-                _editedCustomFeature.parameters.itemById(nonlinearInputDef.id).expression = str(_nonlinearValueInput.value).lower()
-            except:
-                _editedCustomFeature.parameters.add(nonlinearInputDef.id, nonlinearInputDef.name, adsk.core.ValueInput.createByString(str(_nonlinearValueInput.value).lower()), '', True)
-
-            try:
-                _editedCustomFeature.parameters.itemById(nonlinearSizeInputDef.id).expression = _nonlinearSizeValueInput.expression
-            except:
-                _editedCustomFeature.parameters.add(nonlinearSizeInputDef.id, nonlinearSizeInputDef.name, adsk.core.ValueInput.createByString(_nonlinearSizeValueInput.expression), _app.activeProduct.unitsManager.defaultLengthUnits, True)
-
-            try:
-                _editedCustomFeature.parameters.itemById(nonlinearPositionInputDef.id).expression = _nonlinearPositionValueInput.expression
-            except:
-                _editedCustomFeature.parameters.add(nonlinearPositionInputDef.id, nonlinearPositionInputDef.name, adsk.core.ValueInput.createByString(_nonlinearPositionValueInput.expression), '', True)
 
             updateFeature(_editedCustomFeature)
 
@@ -896,6 +910,11 @@ def updateFeature(customFeature: adsk.fusion.CustomFeature) -> bool:
             flipDirection = customFeature.parameters.itemById(flipDirectionInputDef.id).expression.lower() == 'true'
         except:
             flipDirection = False
+
+        try:
+            uniformDistribution = customFeature.parameters.itemById(uniformDistributionInputDef.id).expression.lower() == 'true'
+        except:
+            uniformDistribution = False
         
         absoluteDepthOffset = customFeature.parameters.itemById(absoluteDepthOffsetInputDef.id).value
         relativeDepthOffset = customFeature.parameters.itemById(relativeDepthOffsetInputDef.id).value
@@ -919,7 +938,7 @@ def updateFeature(customFeature: adsk.fusion.CustomFeature) -> bool:
         if curveGeometry is None:
             return True
 
-        pointsAndSizes = calculatePointsAndSizesAlongCurve(curveGeometry, startOffset, endOffset, startSize, endSize, sizeStep, targetGap, flipDirection, nonlinear, nonlinearSize, nonlinearPosition)
+        pointsAndSizes = calculatePointsAndSizesAlongCurve(curveGeometry, startOffset, endOffset, startSize, endSize, sizeStep, targetGap, flipDirection, uniformDistribution, nonlinear, nonlinearSize, nonlinearPosition)
         if len(pointsAndSizes) == 0: return True
 
         if faceEntity.objectType == adsk.fusion.ConstructionPlane.classType():